feat(results): update badges to reflect applied filters and sort options
